chore(exp): Remove legacy external exchanges screen
