+++ src/alert/reducer.ts
@@ -1,12 +1,7 @@
 import { AnyAction } from 'redux'
-import { Actions, ActionTypes, AlertTypes, ShowAlertAction } from 'src/alert/actions'
+import { Actions, ActionTypes, AlertTypes } from 'src/alert/actions'
 import { ErrorMessages } from 'src/app/ErrorMessages'
-import { ALERT_BANNER_DURATION } from 'src/config'
-import {
-  Actions as ExchangeActions,
-  ActionTypes as ExchangeActionTypes,
-} from 'src/exchange/actions'
-import i18n from 'src/i18n'
+import { ActionTypes as ExchangeActionTypes } from 'src/exchange/actions'
 import { RootState } from 'src/redux/reducers'
 
 export enum ErrorDisplayType {
@@ -29,24 +24,11 @@ export type State = Alert | null
 
 const initialState = null
 
-const errorAction = (error: ErrorMessages): ShowAlertAction => ({
-  type: Actions.SHOW,
-  alertType: AlertTypes.ERROR,
-  displayMethod: ErrorDisplayType.BANNER,
-  message: i18n.t(error),
-  dismissAfter: ALERT_BANNER_DURATION,
-  buttonMessage: null,
-  title: null,
-  underlyingError: error,
-})
-
 export const reducer = (
   state: State = initialState,
   action: ActionTypes | ExchangeActionTypes
 ): State => {
   switch (action.type) {
-    case ExchangeActions.WITHDRAW_CELO_FAILED:
-      action = errorAction(action.error)
     case Actions.SHOW:
       return {
         displayMethod: action.displayMethod,
