+++ src/components/AccountNumber.tsx
@@ -3,7 +3,7 @@ import Clipboard from '@react-native-clipboard/clipboard'
 import React from 'react'
 import { useTranslation } from 'react-i18next'
 import { StyleSheet, Text, TouchableOpacity } from 'react-native'
-import { FiatExchangeEvents, HomeEvents } from 'src/analytics/Events'
+import { HomeEvents } from 'src/analytics/Events'
 import ValoraAnalytics from 'src/analytics/ValoraAnalytics'
 import { Screens } from 'src/navigator/Screens'
 import colors from 'src/styles/colors'
@@ -33,10 +33,6 @@ export default function AccountNumber({ address, touchDisabled, location }: Prop
 
     if (location === Screens.TransactionDetailsScreen) {
       ValoraAnalytics.track(HomeEvents.transaction_feed_address_copy)
-    }
-
-    if (location === Screens.ExternalExchanges) {
-      ValoraAnalytics.track(FiatExchangeEvents.cico_cash_out_copy_address)
     }
   }
   // Turns '0xce10ce10ce10ce10ce10ce10ce10ce10ce10ce10'
