+++ src/components/ExchangesBottomSheet.tsx
@@ -3,12 +3,12 @@ import { useTranslation } from 'react-i18next'
 import { StyleSheet, Text, View } from 'react-native'
 import BottomSheetLegacy from 'src/components/BottomSheetLegacy'
 import Touchable from 'src/components/Touchable'
+import { ExternalExchangeProvider } from 'src/exchange/types'
+import { navigate } from 'src/navigator/NavigationService'
+import { Screens } from 'src/navigator/Screens'
 import colors from 'src/styles/colors'
 import fontStyles from 'src/styles/fonts'
 import { Spacing } from 'src/styles/styles'
-import { ExternalExchangeProvider } from 'src/fiatExchanges/ExternalExchanges'
-import { navigate } from 'src/navigator/NavigationService'
-import { Screens } from 'src/navigator/Screens'
 
 interface Props {
   isVisible: boolean
