--- src/exchange/WithdrawCeloQrScannerScreen.tsx
@@ -1,71 +1,0 @@
-// SCREEN that scans QRs and calls |onAddressScanned| param when one is found.
-
-import { NativeStackScreenProps } from '@react-navigation/native-stack'
-import { memoize } from 'lodash'
-import React from 'react'
-import { useDispatch } from 'react-redux'
-import { showError } from 'src/alert/actions'
-import { ErrorMessages } from 'src/app/ErrorMessages'
-import BackButton from 'src/components/BackButton'
-import { ADDRESS_LENGTH } from 'src/exchange/reducer'
-import i18n from 'src/i18n'
-import { nuxNavigationOptions } from 'src/navigator/Headers'
-import { navigateBack } from 'src/navigator/NavigationService'
-import { Screens } from 'src/navigator/Screens'
-import { StackParamList } from 'src/navigator/types'
-import QRScanner from 'src/qrcode/QRScanner'
-import { uriDataFromUrl } from 'src/qrcode/schema'
-import { QrCode } from 'src/send/actions'
-import colors from 'src/styles/colors'
-import fontStyles from 'src/styles/fonts'
-import Logger from 'src/utils/Logger'
-
-const TAG = 'WithdrawCeloQrScannerScreen'
-
-type Props = NativeStackScreenProps<StackParamList, Screens.WithdrawCeloQrScannerScreen>
-
-function fetchAddressFromQrData(data: string): string | null {
-  try {
-    return uriDataFromUrl(data).address
-  } catch (error) {
-    return null
-  }
-}
-
-function WithdrawCeloQrScannerScreen({ route }: Props) {
-  const { onAddressScanned } = route.params
-  const dispatch = useDispatch()
-
-  const onBarCodeDetected = memoize(
-    (qrCode: QrCode) => {
-      const qrData = qrCode.data
-      Logger.debug(TAG, 'Bar code detected: ' + qrData)
-      const address = fetchAddressFromQrData(qrData) || qrData
-      if (address.startsWith('0x') && address.length === ADDRESS_LENGTH) {
-        onAddressScanned(address)
-        navigateBack()
-      } else {
-        dispatch(showError(ErrorMessages.QR_FAILED_INVALID_ADDRESS))
-      }
-    },
-    (qrCode) => qrCode.data
-  )
-
-  return <QRScanner onBarCodeDetected={onBarCodeDetected} />
-}
-
-WithdrawCeloQrScannerScreen.navigationOptions = () => {
-  return {
-    ...nuxNavigationOptions,
-    headerLeft: () => <BackButton color={colors.light} />,
-    headerTitle: i18n.t('withdrawScanQrTitle'),
-    headerTitleStyle: {
-      ...fontStyles.navigationHeader,
-      headerTitleAlign: 'center',
-      fontWeight: undefined,
-      color: colors.light,
-    },
-  }
-}
-
-export default WithdrawCeloQrScannerScreen
