--- src/exchange/WithdrawCeloSummary.test.tsx
@@ -1,27 +1,0 @@
-import { render } from '@testing-library/react-native'
-import BigNumber from 'bignumber.js'
-import * as React from 'react'
-import 'react-native'
-import { Provider } from 'react-redux'
-import WithdrawCeloSummary from 'src/exchange/WithdrawCeloSummary'
-import { createMockStore } from 'test/utils'
-
-const SAMPLE_ADDRESS = '0xcc642068bdbbdeb91f348213492d2a80ab1ed23c'
-const SAMPLE_AMOUNT = new BigNumber(5.001)
-
-const store = createMockStore()
-
-describe('WithdrawCeloSummary', () => {
-  it('renders correctly', () => {
-    const tree = render(
-      <Provider store={store}>
-        <WithdrawCeloSummary
-          amount={SAMPLE_AMOUNT}
-          recipientAddress={SAMPLE_ADDRESS}
-          feeEstimate={new BigNumber(1)}
-        />
-      </Provider>
-    )
-    expect(tree).toMatchSnapshot()
-  })
-})
