+++ src/exchange/types.ts
@@ -1,3 +1,5 @@
+import { Currency } from 'src/utils/currencies'
+
 export interface CeloNewsConfig {
   readMoreUrl?: string
 }
@@ -14,4 +16,11 @@ export interface CeloNewsArticle {
   link: string
   title: string
   type: string
+}
+
+export interface ExternalExchangeProvider {
+  name: string
+  link: string
+  currencies: Currency[]
+  supportedRegions?: string[]
 }
