+++ src/fiatExchanges/ExchangeQR.tsx
@@ -5,17 +5,17 @@ import { useDispatch } from 'react-redux'
 import { FiatExchangeEvents } from 'src/analytics/Events'
 import ValoraAnalytics from 'src/analytics/ValoraAnalytics'
 import BackButton from 'src/components/BackButton'
-import { ExternalExchangeProvider } from 'src/fiatExchanges/ExternalExchanges'
+import { ExternalExchangeProvider } from 'src/exchange/types'
 import i18n from 'src/i18n'
 import Share from 'src/icons/Share'
 import { emptyHeader } from 'src/navigator/Headers'
 import { Screens } from 'src/navigator/Screens'
 import { TopBarIconButton } from 'src/navigator/TopBarButton'
 import { StackParamList } from 'src/navigator/types'
-import { shareQRCode, SVG } from 'src/send/actions'
-import { CICOFlow } from './utils'
 import NewQRCode from 'src/qrcode/NewQRCode'
+import { SVG, shareQRCode } from 'src/send/actions'
 import { QRCodeDataType } from 'src/statsig/types'
+import { CICOFlow } from './utils'
 
 type Props = NativeStackScreenProps<StackParamList, Screens.ExchangeQR>
