+++ src/fiatExchanges/SelectProvider.test.tsx
@@ -3,30 +3,28 @@ import { FetchMock } from 'jest-fetch-mock/types'
 import * as React from 'react'
 import { Provider } from 'react-redux'
 import { MockStoreEnhanced } from 'redux-mock-store'
+import { FiatExchangeEvents } from 'src/analytics/Events'
+import ValoraAnalytics from 'src/analytics/ValoraAnalytics'
 import SelectProviderScreen from 'src/fiatExchanges/SelectProvider'
 import { SelectProviderExchangesLink, SelectProviderExchangesText } from 'src/fiatExchanges/types'
 import { LocalCurrencyCode } from 'src/localCurrency/consts'
 import { navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
 import { getExperimentParams, getFeatureGate } from 'src/statsig'
-import { ExperimentConfigs } from 'src/statsig/constants'
-import { StatsigExperiments } from 'src/statsig/types'
+import { Network } from 'src/transactions/types'
 import { CiCoCurrency, Currency } from 'src/utils/currencies'
 import { createMockStore, getMockStackScreenProps } from 'test/utils'
 import { mockAccount, mockExchanges, mockFiatConnectQuotes, mockProviders } from 'test/values'
 import {
   CICOFlow,
+  LegacyMobileMoneyProvider,
+  PaymentMethod,
   fetchExchanges,
   fetchLegacyMobileMoneyProviders,
   fetchProviders,
   getProviderSelectionAnalyticsData,
-  LegacyMobileMoneyProvider,
-  PaymentMethod,
 } from './utils'
-import ValoraAnalytics from 'src/analytics/ValoraAnalytics'
 import mocked = jest.mocked
-import { FiatExchangeEvents } from 'src/analytics/Events'
-import { Network } from 'src/transactions/types'
 
 const AMOUNT_TO_CASH_IN = 100
 const MOCK_IP_ADDRESS = '1.1.1.7'
@@ -346,11 +344,7 @@ describe(SelectProviderScreen, () => {
       jest.mocked(fetchExchanges).mockResolvedValue(mockExchanges)
     })
 
-    it('renders crypto exchange and navigates to exchanges screen regardless of statsig param for cash outs', async () => {
-      jest.mocked(getExperimentParams).mockReturnValue({
-        addFundsExchangesText: SelectProviderExchangesText.DepositFrom,
-        addFundsExchangesLink: SelectProviderExchangesLink.ExchangeQRScreen,
-      })
+    it('renders for cash outs', async () => {
       const { queryByText, getByText, getByTestId } = render(
         <Provider store={mockStore}>
           <SelectProviderScreen {...mockScreenProps(CICOFlow.CashOut)} />
@@ -363,22 +357,16 @@ describe(SelectProviderScreen, () => {
       expect(queryByText('selectProviderScreen.viewExchanges')).toBeTruthy()
       expect(queryByText('selectProviderScreen.depositFrom')).toBeFalsy()
       expect(queryByText('selectProviderScreen.cryptoExchangeOrWallet')).toBeFalsy()
-      expect(getExperimentParams).not.toHaveBeenCalled()
 
       fireEvent.press(getByText('selectProviderScreen.viewExchanges'))
       expect(navigate).toHaveBeenCalledTimes(1)
-      expect(navigate).toHaveBeenCalledWith(Screens.ExternalExchanges, {
-        currency: Currency.Dollar,
-        isCashIn: false,
+      expect(navigate).toHaveBeenCalledWith(Screens.ExchangeQR, {
+        flow: CICOFlow.CashOut,
         exchanges: mockExchanges,
       })
     })
 
-    it('renders based on params from statsig for cash ins', async () => {
-      jest.mocked(getExperimentParams).mockReturnValue({
-        addFundsExchangesText: SelectProviderExchangesText.DepositFrom,
-        addFundsExchangesLink: SelectProviderExchangesLink.ExchangeQRScreen,
-      })
+    it('renders for cash ins', async () => {
       const { queryByText, getByText, getByTestId } = render(
         <Provider store={mockStore}>
           <SelectProviderScreen {...mockScreenProps()} />
@@ -391,10 +379,6 @@ describe(SelectProviderScreen, () => {
       expect(queryByText('selectProviderScreen.viewExchanges')).toBeFalsy()
       expect(queryByText('selectProviderScreen.depositFrom')).toBeTruthy()
       expect(queryByText('selectProviderScreen.cryptoExchangeOrWallet')).toBeTruthy()
-      expect(getExperimentParams).toHaveBeenCalledTimes(1)
-      expect(getExperimentParams).toHaveBeenCalledWith(
-        ExperimentConfigs[StatsigExperiments.ADD_FUNDS_CRYPTO_EXCHANGE_QR_CODE]
-      )
 
       fireEvent.press(getByText('selectProviderScreen.cryptoExchangeOrWallet'))
       expect(navigate).toHaveBeenCalledTimes(1)
