+++ src/fiatExchanges/utils.tsx
@@ -2,7 +2,7 @@ import firebase from '@react-native-firebase/app'
 import BigNumber from 'bignumber.js'
 import { default as DeviceInfo } from 'react-native-device-info'
 import { FIREBASE_ENABLED } from 'src/config'
-import { ExternalExchangeProvider } from 'src/fiatExchanges/ExternalExchanges'
+import { ExternalExchangeProvider } from 'src/exchange/types'
 import NormalizedQuote from 'src/fiatExchanges/quotes/NormalizedQuote'
 import { ProviderSelectionAnalyticsData } from 'src/fiatExchanges/types'
 import { LocalCurrencyCode } from 'src/localCurrency/consts'
@@ -11,11 +11,11 @@ import { getDynamicConfigParams } from 'src/statsig'
 import { DynamicConfigs } from 'src/statsig/constants'
 import { StatsigDynamicConfigs } from 'src/statsig/types'
 import { TokenBalance } from 'src/tokens/slice'
+import { Network } from 'src/transactions/types'
+import Logger from 'src/utils/Logger'
 import { CiCoCurrency } from 'src/utils/currencies'
 import { fetchWithTimeout } from 'src/utils/fetchWithTimeout'
-import Logger from 'src/utils/Logger'
 import networkConfig from 'src/web3/networkConfig'
-import { Network } from 'src/transactions/types'
 
 const TAG = 'fiatExchanges:utils'
