+++ src/navigator/Navigator.tsx
@@ -4,8 +4,8 @@ import {
   useBottomSheetDynamicSnapPoints,
 } from '@gorhom/bottom-sheet'
 import {
-  createNativeStackNavigator,
   NativeStackNavigationOptions,
+  createNativeStackNavigator,
 } from '@react-navigation/native-stack'
 import { createBottomSheetNavigator } from '@th3rdwave/react-navigation-bottom-sheet'
 import * as React from 'react'
@@ -30,29 +30,14 @@ import BackupQuiz, { navOptionsForQuiz } from 'src/backup/BackupQuiz'
 import ConsumerIncentivesHomeScreen from 'src/consumerIncentives/ConsumerIncentivesHomeScreen'
 import DappKitAccountScreen from 'src/dappkit/DappKitAccountScreen'
 import DappKitSignTxScreen from 'src/dappkit/DappKitSignTxScreen'
-import DappShortcutsRewards from 'src/dapps/DappShortcutsRewards'
 import DappShortcutTransactionRequest from 'src/dapps/DappShortcutTransactionRequest'
+import DappShortcutsRewards from 'src/dapps/DappShortcutsRewards'
 import EscrowedPaymentListScreen from 'src/escrow/EscrowedPaymentListScreen'
 import ReclaimPaymentConfirmationScreen from 'src/escrow/ReclaimPaymentConfirmationScreen'
-import WithdrawCeloQrScannerScreen from 'src/exchange/WithdrawCeloQrScannerScreen'
-import WithdrawCeloReviewScreen from 'src/exchange/WithdrawCeloReviewScreen'
-import WithdrawCeloScreen from 'src/exchange/WithdrawCeloScreen'
-import FiatDetailsScreen from 'src/fiatconnect/FiatDetailsScreen'
-import KycDenied from 'src/fiatconnect/kyc/KycDenied'
-import KycExpired from 'src/fiatconnect/kyc/KycExpired'
-import KycPending from 'src/fiatconnect/kyc/KycPending'
-import KycLanding from 'src/fiatconnect/KycLanding'
-import FiatConnectLinkAccountScreen from 'src/fiatconnect/LinkAccountScreen'
-import FiatConnectRefetchQuoteScreen from 'src/fiatconnect/RefetchQuoteScreen'
-import FiatConnectReviewScreen from 'src/fiatconnect/ReviewScreen'
-import FiatConnectTransferStatusScreen from 'src/fiatconnect/TransferStatusScreen'
 import BidaliScreen from 'src/fiatExchanges/BidaliScreen'
 import CashInSuccess from 'src/fiatExchanges/CashInSuccess'
 import CoinbasePayScreen from 'src/fiatExchanges/CoinbasePayScreen'
 import ExchangeQR from 'src/fiatExchanges/ExchangeQR'
-import ExternalExchanges, {
-  externalExchangesScreenOptions,
-} from 'src/fiatExchanges/ExternalExchanges'
 import FiatExchangeAmount from 'src/fiatExchanges/FiatExchangeAmount'
 import FiatExchangeCurrency, {
   fiatExchangesOptionsScreenOptions,
@@ -61,14 +46,23 @@ import SelectProviderScreen from 'src/fiatExchanges/SelectProvider'
 import SimplexScreen from 'src/fiatExchanges/SimplexScreen'
 import Spend, { spendScreenOptions } from 'src/fiatExchanges/Spend'
 import WithdrawSpend from 'src/fiatExchanges/WithdrawSpend'
+import FiatDetailsScreen from 'src/fiatconnect/FiatDetailsScreen'
+import KycLanding from 'src/fiatconnect/KycLanding'
+import FiatConnectLinkAccountScreen from 'src/fiatconnect/LinkAccountScreen'
+import FiatConnectRefetchQuoteScreen from 'src/fiatconnect/RefetchQuoteScreen'
+import FiatConnectReviewScreen from 'src/fiatconnect/ReviewScreen'
+import FiatConnectTransferStatusScreen from 'src/fiatconnect/TransferStatusScreen'
+import KycDenied from 'src/fiatconnect/kyc/KycDenied'
+import KycExpired from 'src/fiatconnect/kyc/KycExpired'
+import KycPending from 'src/fiatconnect/kyc/KycPending'
 import NotificationCenter from 'src/home/NotificationCenter'
 import { currentLanguageSelector } from 'src/i18n/selectors'
 import PhoneNumberLookupQuotaScreen from 'src/identity/PhoneNumberLookupQuotaScreen'
 import ImportWallet from 'src/import/ImportWallet'
 import KeylessBackupPhoneCodeInput from 'src/keylessBackup/KeylessBackupPhoneCodeInput'
 import KeylessBackupPhoneInput from 'src/keylessBackup/KeylessBackupPhoneInput'
-import SetUpKeylessBackup from 'src/keylessBackup/SetUpKeylessBackup'
 import KeylessBackupProgress from 'src/keylessBackup/KeylessBackupProgress'
+import SetUpKeylessBackup from 'src/keylessBackup/SetUpKeylessBackup'
 import SignInWithEmail from 'src/keylessBackup/SignInWithEmail'
 import WalletSecurityPrimer from 'src/keylessBackup/WalletSecurityPrimer'
 import Language from 'src/language/Language'
@@ -83,9 +77,9 @@ import {
   noHeaderGestureDisabled,
   nuxNavigationOptions,
 } from 'src/navigator/Headers'
-import { getInitialRoute } from 'src/navigator/initialRoute'
 import QRNavigator from 'src/navigator/QRNavigator'
 import { Screens } from 'src/navigator/Screens'
+import { getInitialRoute } from 'src/navigator/initialRoute'
 import { StackParamList } from 'src/navigator/types'
 import NftsInfoCarousel from 'src/nfts/NftsInfoCarousel'
 import ChooseYourAdventure from 'src/onboarding/ChooseYourAdventure'
@@ -295,26 +289,6 @@ const sendScreens = (Navigator: typeof Stack) => (
   </>
 )
 
-const exchangeScreens = (Navigator: typeof Stack) => (
-  <>
-    <Navigator.Screen
-      name={Screens.WithdrawCeloScreen}
-      component={WithdrawCeloScreen}
-      options={WithdrawCeloScreen.navigationOptions}
-    />
-    <Navigator.Screen
-      name={Screens.WithdrawCeloQrScannerScreen}
-      component={WithdrawCeloQrScannerScreen}
-      options={WithdrawCeloQrScannerScreen.navigationOptions}
-    />
-    <Navigator.Screen
-      name={Screens.WithdrawCeloReviewScreen}
-      component={WithdrawCeloReviewScreen}
-      options={WithdrawCeloReviewScreen.navigationOptions}
-    />
-  </>
-)
-
 const consumerIncentivesScreens = (Navigator: typeof Stack) => (
   <>
     <Navigator.Screen
@@ -397,11 +371,6 @@ const settingsScreens = (Navigator: typeof Stack) => (
       options={headerWithBackButton}
       name={Screens.SupportContact}
       component={SupportContact}
-    />
-    <Navigator.Screen
-      options={externalExchangesScreenOptions}
-      name={Screens.ExternalExchanges}
-      component={ExternalExchanges}
     />
     <Navigator.Screen options={spendScreenOptions} name={Screens.Spend} component={Spend} />
     <Navigator.Screen
@@ -640,7 +609,6 @@ export function MainStackScreen() {
       {sendScreens(Stack)}
       {nuxScreens(Stack)}
       {verificationScreens(Stack)}
-      {exchangeScreens(Stack)}
       {backupScreens(Stack)}
       {consumerIncentivesScreens(Stack)}
       {settingsScreens(Stack)}
