+++ src/navigator/QRNavigator.tsx
@@ -14,7 +14,7 @@ import { ScrollPager } from 'react-native-tab-view'
 import { useDispatch, useSelector } from 'react-redux'
 import { QrScreenEvents } from 'src/analytics/Events'
 import ValoraAnalytics from 'src/analytics/ValoraAnalytics'
-import { ExternalExchangeProvider } from 'src/fiatExchanges/ExternalExchanges'
+import { ExternalExchangeProvider } from 'src/exchange/types'
 import { fetchExchanges } from 'src/fiatExchanges/utils'
 import { noHeader } from 'src/navigator/Headers'
 import { Screens } from 'src/navigator/Screens'
