+++ src/qrcode/NewQRCode.tsx
@@ -6,7 +6,7 @@ import { useSelector } from 'react-redux'
 import { nameSelector } from 'src/account/selectors'
 import Button from 'src/components/Button'
 import ExchangesBottomSheet from 'src/components/ExchangesBottomSheet'
-import { ExternalExchangeProvider } from 'src/fiatExchanges/ExternalExchanges'
+import { ExternalExchangeProvider } from 'src/exchange/types'
 import Paste from 'src/icons/Paste'
 import StyledQRCode from 'src/qrcode/StyledQRCode'
 import { SVG } from 'src/send/actions'
