+++ src/statsig/constants.ts
@@ -1,4 +1,3 @@
-import { SelectProviderExchangesLink, SelectProviderExchangesText } from 'src/fiatExchanges/types'
 import {
   QRCodeDataType,
   QRCodeStyle,
@@ -41,13 +40,6 @@ export const FeatureGates = {
 
 export const ExperimentConfigs = {
   // NOTE: the keys of defaultValues MUST be parameter names
-  [StatsigExperiments.ADD_FUNDS_CRYPTO_EXCHANGE_QR_CODE]: {
-    experimentName: StatsigExperiments.ADD_FUNDS_CRYPTO_EXCHANGE_QR_CODE,
-    defaultValues: {
-      addFundsExchangesText: SelectProviderExchangesText.CryptoExchange,
-      addFundsExchangesLink: SelectProviderExchangesLink.ExternalExchangesScreen,
-    },
-  },
   [StatsigExperiments.CHOOSE_YOUR_ADVENTURE]: {
     experimentName: StatsigExperiments.CHOOSE_YOUR_ADVENTURE,
     defaultValues: {
